requirements.txt
# Core Django
Django==5.2.7
python-dotenv==1.2.1

# PDF Generation
weasyprint==66.0
Pillow==12.0.0

# Google Drive API
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
